♻️  Refactor: modificación de entidades y repositorios

+++ src/main/java/zuzzok/arecorecicla/data/repositories/BeneficioRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import zuzzok.arecorecicla.data.models.Beneficio;
 
-public interface BeneficioRepository extends JpaRepository<Beneficio, Integer> {
+public interface BeneficioRepository extends JpaRepository<Beneficio, Long> {
 
 }
