+++ src/main/java/zuzzok/arecorecicla/data/repositories/CuponRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import zuzzok.arecorecicla.data.models.Cupon;
 
-public interface CuponRepository extends JpaRepository<Cupon, Integer> {
+public interface CuponRepository extends JpaRepository<Cupon, Long> {
 
 }
