+++ src/main/java/zuzzok/arecorecicla/data/repositories/EcopuntoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import zuzzok.arecorecicla.data.models.Ecopunto;
 
-public interface EcopuntoRepository extends JpaRepository<Ecopunto, Integer> {
+public interface EcopuntoRepository extends JpaRepository<Ecopunto, Long> {
 
 }
