+++ src/main/java/zuzzok/arecorecicla/data/repositories/GeneroRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import zuzzok.arecorecicla.data.models.Genero;
 
-public interface GeneroRepository extends JpaRepository<Genero, Integer> {
+public interface GeneroRepository extends JpaRepository<Genero, Long> {
 
 }
