+++ src/main/java/zuzzok/arecorecicla/data/repositories/PuntoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import zuzzok.arecorecicla.data.models.Punto;
 
-public interface PuntoRepository extends JpaRepository<Punto, Integer> {
+public interface PuntoRepository extends JpaRepository<Punto, Long> {
 
 }
