+++ src/main/java/zuzzok/arecorecicla/data/repositories/ReciclableRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import zuzzok.arecorecicla.data.models.Reciclable;
 
-public interface ReciclableRepository extends JpaRepository<Reciclable, Integer> {
+public interface ReciclableRepository extends JpaRepository<Reciclable, Long> {
 
 }
